+++ apps/frontend/src/libs/enums/query-parameter-name.enum.ts
@@ -1,6 +1,7 @@
 const QueryParameterName = {
 	PAGE: "page",
 	PAGE_SIZE: "pageSize",
+	SEARCH: "q",
 } as const;
 
 export { QueryParameterName };